requirements.txt
langchain
langchain_mistralai
chromadb
flask
Gitpython
python-dotenv
langchain-community